♻️ refactor(player): finalize TransportControls colocation and public PlayButton alias
`AudioPlayer.TransportControls` was introduced in this unreleased v2.2.0
cycle as a rename of the original `AudioPlayer.PlayButton` compound.
Since it never shipped publicly, keeping it as a static member plus a
deprecated `PlayButton` alias was pure internal churn.

Finalize the surface so that the internal component stays named
`TransportControls` (reflects the role — Prev + Play + Next group) while
the single public static member is `AudioPlayer.PlayButton` (matches
`activeUI.playButton` and preserves the only name consumers ever saw):

- fix relative imports inside the TransportControls folder
  (`../Icon` → `../../Icon`) after the earlier file move
- drop the deprecated `PlayButton` / `PlayButtonProps` re-exports from
  `Controller/Button/index.ts`
- remove the `AudioPlayer.TransportControls` static member and the
  `@deprecated PlayButton` entry in `AudioPlayerCompound`; a short
  comment on the type declaration notes the public-to-internal mapping
  and the planned v3 public rename
- point `Button/__tests__/PlayBtn.a11y.test.tsx` at the new
  `TransportControls/PlayBtn` path
- rewrite the compound-slots identity check: remove
  `expect(AudioPlayer.TransportControls).toBe(AudioPlayer.PlayButton)`
  and the `TransportControls` defined-check, keeping only the
  consumer-visible members

Also clean up the user-facing docs accordingly:
- CHANGELOG: drop the `### ⚠️ Deprecations` section (nothing was
  publicly deprecated), compound slot table lists a single
  `AudioPlayer.PlayButton` row
- README: matching table update

A follow-up entry in `.claude/docs/v3-deferred.md` tracks the eventual
public rename to `TransportControls` (plus the `activeUI.playButton` /
`templateArea.playButton` keys) for v3.
