+++ package/src/components/AudioPlayer/Interface/Controller/Button/TransportControls/PlayBtn.tsx
@@ -6,7 +6,7 @@ import { usePlaybackContext } from "@/hooks/context/usePlaybackContext";
 import { useResourceContext } from "@/hooks/context/useResourceContext";
 import { StyledBtn } from "@/ui/StyledBtn";
 import { MdPauseCircleFilled, MdPlayCircleFilled } from "@/components/icons";
-import { Icon } from "../Icon";
+import { Icon } from "../../Icon";
 
 export const PlayBtn: FC = memo(function PlayBtn() {
   const { isPlaying } = usePlaybackContext();
