+++ package/src/components/AudioPlayer/Interface/Controller/Button/index.ts
@@ -1,11 +1,6 @@
-export * from "./PlayBtn";
-export * from "./PrevBtn";
-export * from "./NextBtn";
 export * from "./TransportControls";
+export * from "./TransportControls/PlayBtn";
+export * from "./TransportControls/PrevBtn";
+export * from "./TransportControls/NextBtn";
 export * from "./RepeatTypeBtn";
 export * from "./VolumeIcon";
-
-/** @deprecated Use `TransportControls` instead. Will be removed in the next major version. */
-export { TransportControls as PlayButton } from "./TransportControls";
-/** @deprecated Use `TransportControlsProps` instead. Will be removed in the next major version. */
-export type { TransportControlsProps as PlayButtonProps } from "./TransportControls";
